+++ src/main/java/misterpemodder/tmo/main/client/render/TileEntityTitaniumAnvilRenderer.java
@@ -47,7 +47,7 @@ public class TileEntityTitaniumAnvilRenderer extends TileEntitySpecialRenderer<T
             	j = (int) facing.getHorizontalAngle();
             }
                 
-            GlStateManager.rotate((float)j, 0.0F, 1.0F, 0.0F);
+            GlStateManager.rotate(j, 0.0F, 1.0F, 0.0F);
             GlStateManager.translate(-0.5F, -0.5F, -0.5F);
         	
             GlStateManager.disableLighting();
