+++ src/main/java/misterpemodder/tmo/main/client/render/TileEntityTitaniumChestRenderer.java
@@ -64,7 +64,7 @@ public class TileEntityTitaniumChestRenderer extends TileEntitySpecialRenderer<T
         	j = (int) facing.getHorizontalAngle();
         }
             
-        GlStateManager.rotate((float)j, 0.0F, 1.0F, 0.0F);
+        GlStateManager.rotate(j, 0.0F, 1.0F, 0.0F);
         GlStateManager.translate(-0.5F, -0.5F, -0.5F);
         float f = te.prevLidAngle + (te.lidAngle - te.prevLidAngle) * partialTicks;
