+++ src/main/java/misterpemodder/tmo/main/utils/StringUtils.java
@@ -11,7 +11,7 @@ import net.minecraft.client.gui.FontRenderer;
 public class StringUtils {
 	
 	public static void drawCenteredString(FontRenderer fontRenderer, String text, int x, int y) {
-        fontRenderer.drawStringWithShadow(text, (float)(x - fontRenderer.getStringWidth(text) / 2), (float)y, Color.WHITE.getRGB());
+        fontRenderer.drawStringWithShadow(text, x - fontRenderer.getStringWidth(text) / 2, y, Color.WHITE.getRGB());
     }
 	
 	public static List<String> parseTooltip(String line) {
